+++ orix/plot/crystal_map_plot.py
@@ -20,16 +20,19 @@
 from typing import TYPE_CHECKING
 
 from matplotlib.axes import Axes
-import matplotlib.colorbar as mbar
+import matplotlib.colorbar as mcolorbar
 from matplotlib.image import AxesImage
 import matplotlib.patches as mpatches
 from matplotlib.projections import register_projection
-from matplotlib_scalebar.dimension import _Dimension
-from matplotlib_scalebar.scalebar import ScaleBar
 from mpl_toolkits.axes_grid1 import make_axes_locatable
 import numpy as np
 
+from orix.constants import installed, verify_dependency_or_raise
+
 if TYPE_CHECKING:  # pragma: no cover
+    if installed["matplotlib-scalebar"]:
+        from matplotlib_scalebar.scalebar import ScaleBar
+
     from orix.crystal_map.crystal_map import CrystalMap
 
 
@@ -37,16 +40,16 @@ class CrystalMapPlot(Axes):
     """Plotting of a :class:`~orix.crystal_map.crystal_map.CrystalMap`."""
 
     name = "plot_map"
-    _data_axes = None
-    _data_shape = None
-    colorbar = None
-    scalebar = None
+    _data_axes: tuple[int, ...] | None = None
+    _data_shape: tuple[int, ...] | None = None
+    colorbar: mcolorbar.Colorbar | None = None
+    scalebar: "ScaleBar | None" = None
 
     def plot_map(
         self,
         crystal_map: "CrystalMap",
         value: np.ndarray | None = None,
-        scalebar: bool = True,
+        scalebar: bool | None = None,
         scalebar_properties: dict | None = None,
         legend: bool = True,
         legend_properties: dict | None = None,
@@ -66,8 +69,9 @@ class CrystalMapPlot(Axes):
             Attribute array to plot. If value is ``None`` (default), a
             phase map is plotted.
         scalebar
-            Whether to add a scalebar (default is ``True``) along the
-            horizontal map dimension.
+            Whether to add a scalebar along the horizontal map
+            dimension. If not given, a scalebar is added if
+            :mod:`matplotlib-scalebar` is installed.
         scalebar_properties
             Dictionary of keyword arguments passed to
             :class:`mpl_toolkits.axes_grid1.anchored_artists.AnchoredSizeBar`.
@@ -147,6 +151,14 @@ class CrystalMapPlot(Axes):
 
         >>> #plt.imsave("image2.png", im.get_array())
         """
+        # Add scalebar here to potentially raise an error early
+        if scalebar is None:
+            scalebar = installed["matplotlib-scalebar"]
+        if scalebar:
+            if scalebar_properties is None:
+                scalebar_properties = {}
+            _ = self.add_scalebar(crystal_map, **scalebar_properties)
+
         self._data_shape = crystal_map._original_shape
 
         patches = None
@@ -177,19 +189,13 @@ class CrystalMapPlot(Axes):
                 legend_properties = {}
             self._add_legend(patches, **legend_properties)
 
-        # Scalebar
-        if scalebar:
-            if scalebar_properties is None:
-                scalebar_properties = {}
-            _ = self.add_scalebar(crystal_map, **scalebar_properties)
-
         im = self.imshow(X=data, **kwargs)
         if override_status_bar:
             im = self._override_status_bar(im, crystal_map)
 
         return im
 
-    def add_scalebar(self, crystal_map: "CrystalMap", **kwargs) -> ScaleBar:
+    def add_scalebar(self, crystal_map: "CrystalMap", **kwargs) -> "ScaleBar":
         """Add a scalebar to the axes instance and return it.
 
         The scalebar is also available as an attribute :attr:`scalebar`.
@@ -199,8 +205,8 @@ class CrystalMapPlot(Axes):
         crystal_map
             Crystal map instance to obtain necessary data from.
         **kwargs
-            Keyword arguments passed to
-            :class:`matplotlib_scalebar.scalebar.ScaleBar`.
+            Keyword arguments passed on to
+            :class:`~matplotlib_scalebar.scalebar.ScaleBar`.
 
         Returns
         -------
@@ -218,7 +224,18 @@ class CrystalMapPlot(Axes):
         >>> ax = fig.add_subplot(projection="plot_map")
         >>> im = ax.plot_map(xmap, scalebar=False)
         >>> sbar = ax.add_scalebar(xmap, location=4, frameon=False)
+
+        Notes
+        -----
+        This function requires `matplotlib-scalebar` to be installed.
+        Please see the :doc:`installation guide <user/installation>`
+        for details.
         """
+        verify_dependency_or_raise("matplotlib-scalebar", "Adding a scalebar")
+
+        from matplotlib_scalebar.dimension import _Dimension
+        from matplotlib_scalebar.scalebar import ScaleBar
+
         # Get whether z, y or x
         last_axis = crystal_map.ndim - 1
         horizontal = crystal_map._coordinate_axes[last_axis]
@@ -232,7 +249,8 @@ class CrystalMapPlot(Axes):
         else:
             dim = _Dimension(scan_unit)
 
-        # Set up arguments to AnchoredSizeBar() if not already present in kwargs
+        # Set up arguments to AnchoredSizeBar() if not already present
+        # in kwargs
         d = dict(
             pad=0.2,
             sep=3,
@@ -243,7 +261,6 @@ class CrystalMapPlot(Axes):
         )
         [kwargs.setdefault(k, v) for k, v in d.items()]
 
-        # Create scalebar
         bar = ScaleBar(
             dx=crystal_map._step_sizes[horizontal],
             units=crystal_map.scan_unit,
@@ -298,7 +315,7 @@ class CrystalMapPlot(Axes):
 
         image.set_data(image_data)
 
-    def add_colorbar(self, label: str | None = None, **kwargs) -> mbar.Colorbar:
+    def add_colorbar(self, label: str | None = None, **kwargs) -> mcolorbar.Colorbar:
         """Add an opinionated colorbar to the figure and return it.
 
         The colorbar is also available as an attribute :attr:`colorbar`.