+++ orix/tests/test_crystal_map/test_crystal_map.py
@@ -18,18 +18,23 @@
 #
 
 import matplotlib.pyplot as plt
-from matplotlib_scalebar.scalebar import ScaleBar
 import numpy as np
 import pytest
 
-from orix.crystal_map import CrystalMap, Phase, PhaseList, create_coordinate_arrays
+from orix.constants import installed
+from orix.crystal_map import (
+    CrystalMap,
+    Phase,
+    PhaseList,
+    create_coordinate_arrays,
+)
 from orix.crystal_map.crystal_map import _data_slices_from_coordinates
 from orix.plot import CrystalMapPlot
 from orix.quaternion import Orientation, Rotation
 from orix.quaternion.symmetry import C2, C3, C4, O
 
-# Note that many parts of the CrystalMap() class are tested while
-# testing IO and the Phase() and PhaseList() classes
+# Note that many parts of the crystal map class are tested while testing
+# IO and the phase classes
 
 
 class TestCrystalMap:
@@ -209,8 +214,18 @@ class TestCrystalMap:
         "crystal_map_input, phase_names, phase_ids, desired_phase_names",
         [
             (((7, 4), (1, 1), 1, [0]), ["a", "b", "c"], [0, 1, 2], ["a"]),
-            (((7, 4), (1, 1), 1, [0, 1]), ["a", "b", "c"], [0, 2, 1], ["a", "c"]),
-            (((7, 4), (1, 1), 1, [0, 2]), ["a", "b", "c"], [0, 2, 1], ["a", "b"]),
+            (
+                ((7, 4), (1, 1), 1, [0, 1]),
+                ["a", "b", "c"],
+                [0, 2, 1],
+                ["a", "c"],
+            ),
+            (
+                ((7, 4), (1, 1), 1, [0, 2]),
+                ["a", "b", "c"],
+                [0, 2, 1],
+                ["a", "b"],
+            ),
             (((7, 4), (1, 1), 1, [3]), ["a", "b", "c"], [0, 2, 1], ["a"]),
         ],
         indirect=["crystal_map_input"],
@@ -1042,7 +1057,12 @@ class TestCrystalMapShape:
         indirect=["crystal_map_input"],
     )
     def test_data_slice_from_coordinates_masked(
-        self, crystal_map_input, slices, expected_size, expected_shape, expected_slices
+        self,
+        crystal_map_input,
+        slices,
+        expected_size,
+        expected_shape,
+        expected_slices,
     ):
         xmap = CrystalMap(**crystal_map_input)
 
@@ -1108,14 +1128,27 @@ class TestCrystalMapPlotMethod:
         fig1 = xmap.plot(return_figure=True)
         assert isinstance(fig1.axes[0], CrystalMapPlot)
 
-        sbar1 = fig1.axes[0].artists[0]
-        assert isinstance(sbar1, ScaleBar)
-        assert sbar1.dimension.base_units == xmap.scan_unit
-        assert sbar1.location == 3  # "lower left"
-
         prop = np.arange(xmap.size)
+        fig2 = xmap.plot(return_figure=True, remove_padding=True, overlay=prop)
+        ax2 = fig2.axes[0]
+
+        # One effect of "removing padding"
+        assert ax2._xmargin == 0
+        assert ax2._ymargin == 0
+
+        plt.close("all")
+
+    @pytest.mark.skipif(
+        not installed["matplotlib-scalebar"], reason="Requires matplotlib-scalebar"
+    )
+    def test_plot_with_scalebar(self, crystal_map):
+        from matplotlib_scalebar.scalebar import ScaleBar
+
+        fig1 = crystal_map.plot(return_figure=True)
+
+        prop = np.arange(crystal_map.size)
         location = "upper right"
-        fig2 = xmap.plot(
+        fig2 = crystal_map.plot(
             return_figure=True,
             remove_padding=True,
             overlay=prop,
@@ -1127,6 +1160,10 @@ class TestCrystalMapPlotMethod:
         assert ax2._xmargin == 0
         assert ax2._ymargin == 0
 
+        sbar1 = fig1.axes[0].artists[0]
+        assert isinstance(sbar1, ScaleBar)
+        assert sbar1.dimension.base_units == crystal_map.scan_unit
+        assert sbar1.location == 3  # "lower left"
         sbar2 = fig2.axes[0].artists[0]
         assert sbar2.location == 1  # "upper right"